# [FIX] res_currency: add order by to _select_companies_rates query