@odony 2022-08-24
Nice work, thanks!

Minor: some useful information seems to be missing from the commit message (timings, reason why the fix was chosen, etc.), while being present in the PR desc (which is great, btw). Would you mind updating it?


I have also 2 small questions, out of curiosity:
- with regard to the different plans in PG10 vs PG12, have you substantiated that difference, e.g. with a specific change upstream that could explain it, instead of the usual reasons (work_mem, cardinality, statistics, etc.)
- could you elaborate a bit on why adding the extra ORDER BY modifies the handling of the JOIN in this specific case?

Thanks!